Re-factored pages
