+++ src/App.js
@@ -2,221 +2,126 @@ import { useState } from "react";
 import eventsData from "./data";
 import { v1 as generateUniqueID } from "uuid";
 // import Attendees from "./Attendees";
-// import Event from "./Components/Event";
-// import Footer from "./Components/Footer";
-// import Header from "./Components/Header";
-// import NewEventForm from "./Components/NewEventForm";
+import Event from "./Components/Event";
+import Footer from "./Components/Footer";
+import Header from "./Components/Header";
+import NewEventForm from "./Components/NewEventForm";
 
 function App() {
-  const [events, setEvents] = useState(eventsData);
+    const [events, setEvents] = useState(eventsData);
 
-  const [showAttendees, setShowAttendees] = useState(false);
+    const [selectOption, setSelectOption] = useState("");
 
-  const [selectOption, setSelectOption] = useState("");
-
-  const [newEvent, setNewEvent] = useState({
-    id: "",
-    eventType: "",
-    name: "",
-    organizer: "",
-    eventImage: "",
-    date: "",
-    people: [],
-  });
-
-  function addEvent() {
-    const createEvent = {
-      id: generateUniqueID(),
-      eventType: selectOption,
-      name: newEvent.name,
-      organizer: newEvent.organizer,
-      eventImage: newEvent.eventImage || "https://loremflickr.com/640/480/",
-      date: newEvent.date,
-      people: [],
-    };
-    handleAddEvent(createEvent);
-  }
-
-  function handleSelectChange(e) {
-    setSelectOption(e.target.value);
-  }
-
-  function handleSubmit(e) {
-    e.preventDefault();
-    addEvent();
-    resetEventForm();
-  }
-
-  function handleTextChange(e) {
-    setNewEvent({
-      ...newEvent,
-      [e.target.id]: e.target.value,
+    const [newEvent, setNewEvent] = useState({
+        id: "",
+        eventType: "",
+        name: "",
+        organizer: "",
+        eventImage: "",
+        date: "",
+        people: [],
     });
-  }
 
-  function resetEventForm() {
-    setNewEvent({
-      id: "",
-      eventType: "",
-      name: "",
-      organizer: "",
-      eventImage: "",
-      date: "",
-    });
-    setSelectOption("");
-  }
+    function addEvent() {
+        const createEvent = {
+            id: generateUniqueID(),
+            eventType: selectOption,
+            name: newEvent.name,
+            organizer: newEvent.organizer,
+            eventImage:
+                newEvent.eventImage || "https://loremflickr.com/640/480/",
+            date: newEvent.date,
+            people: [],
+        };
+        handleAddEvent(createEvent);
+    }
 
-  function handleAddEvent(event) {
-    setEvents([event, ...events]);
-  }
+    function handleSelectChange(e) {
+        setSelectOption(e.target.value);
+    }
 
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
-  }
+    function handleSubmit(e) {
+        e.preventDefault();
+        addEvent();
+        resetEventForm();
+    }
 
-  function updateEventAttendance(eventId, attendeeId) {
-    const eventArray = [...events];
-    const eventIndex = eventArray.findIndex((event) => eventId === event.id);
-    const event = { ...eventArray[eventIndex] };
-    const personIndex = event.people.findIndex(
-      (person) => person.id === attendeeId
-    );
-    const peopleArray = [...event.people];
-    peopleArray[personIndex].attendance = !peopleArray[personIndex].attendance;
-    event.people = peopleArray;
-    eventArray[eventIndex] = event;
-    setEvents(eventArray);
-  }
+    function handleTextChange(e) {
+        setNewEvent({
+            ...newEvent,
+            [e.target.id]: e.target.value,
+        });
+    }
 
-  return (
-    <div className="App">
-      <>
-        <header>
-          <h1 className="color-change-5x">RSVP App</h1>
-        </header>
-      </>
-      <main>
-        <div className="new-event">
-          <>
-            <form onSubmit={handleSubmit}>
-              <h3>Create a new event</h3>
-              <label htmlFor="name">Event name:</label>
-              <input
-                type="text"
-                id="name"
-                onChange={handleTextChange}
-                value={newEvent.name}
-              />
+    function resetEventForm() {
+        setNewEvent({
+            id: "",
+            eventType: "",
+            name: "",
+            organizer: "",
+            eventImage: "",
+            date: "",
+        });
+        setSelectOption("");
+    }
 
-              <label htmlFor="organizer">Organizer:</label>
-              <input
-                type="text"
-                id="organizer"
-                onChange={handleTextChange}
-                value={newEvent.organizer}
-              />
+    function handleAddEvent(event) {
+        setEvents([event, ...events]);
+    }
 
-              <label htmlFor="eventImage">Event image:</label>
-              <input
-                type="text"
-                id="eventImage"
-                onChange={handleTextChange}
-                value={newEvent.eventImage}
-              />
-              <label htmlFor="eventType">Event type:</label>
-              <select id="eventType" onChange={handleSelectChange}>
-                <option value=""></option>
-                <option value="Birthday">Birthday</option>
-                <option value="Anniversary">Anniversary</option>
-                <option value="Intramural Sport">Intramural Sport</option>
-                <option value="Watch Party">Watch Party</option>
-                <option value="wedding">Wedding</option>
-              </select>
-              <br />
-              <input type="submit" />
-            </form>
-          </>
-        </div>
-        <div className="events">
-          <ul>
-            {events.map((event) => {
-              const { people: attendees } = event;
+    function updateEventAttendance(eventId, attendeeId) {
+        const eventArray = [...events];
+        const eventIndex = eventArray.findIndex(
+            (event) => eventId === event.id
+        );
+        const event = { ...eventArray[eventIndex] };
+        const personIndex = event.people.findIndex(
+            (person) => person.id === attendeeId
+        );
+        const peopleArray = [...event.people];
+        peopleArray[personIndex].attendance =
+            !peopleArray[personIndex].attendance;
+        event.people = peopleArray;
+        eventArray[eventIndex] = event;
+        setEvents(eventArray);
+    }
 
-              return (
-                <>
-                  <li key={event.id}>
-                    <img src={event.eventImage} alt={event.name} />
-                    <h5>
-                      {event.name} {event.eventType}
-                    </h5>
-                    <br />
-                    <span>Organized by: {event.organizer} </span>
-                    <br />
-                    <>
-                      <button onClick={toggleEventAttendees}>
-                        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
-                      </button>
+    return (
+        <div className="App">
+            <Header />
+            <main>
+                <div className="new-event">
+                    <NewEventForm
+                        handleTextChange={handleTextChange}
+                        newEvent={newEvent}
+                        handleSubmit={handleSubmit}
+                        handleSelectChange={handleSelectChange}
+                    />
+                </div>
+                <div className="events">
+                    <ul>
+                        {events.map((event) => {
+                            const { people: attendees } = event;
 
-                      {showAttendees ? (
-                        <div className="attendees">
-                          {attendees.map((attendee, index) => (
-                            <>
-                              <div key={attendee.id} className="attendee">
-                                <p>
-                                  <img
-                                    src={attendee.avatar}
-                                    alt={attendee.firstName}
-                                  />
-                                  {"   "}
-                                  <span>
-                                    {" "}
-                                    {attendee.firstName} {attendee.lastName}{" "}
-                                  </span>
-                                </p>
-                                <p>
-                                  <button
-                                    className="clickable"
-                                    onClick={() =>
-                                      updateEventAttendance(
-                                        event.id,
-                                        attendee.id
-                                      )
+                            return (
+                                <Event
+                                    key={event.id}
+                                    event={event}
+                                    attendees={attendees}
+                                    updateEventAttendance={
+                                        updateEventAttendance
                                     }
-                                  >
-                                    Attending:
-                                  </button>
-                                  <span>
-                                    {attendee.attendance ? "✅" : "❌"}
-                                  </span>
-                                </p>
-
-                                <p>
-                                  <span>Note:</span> {attendee.note}
-                                </p>
-                              </div>
-                            </>
-                          ))}
-                        </div>
-                      ) : null}
-                    </>
-                  </li>
-                </>
-              );
-            })}
-          </ul>
+                                />
+                            );
+                        })}
+                    </ul>
+                </div>
+            </main>
+            <>
+                <Footer />
+            </>
         </div>
-      </main>
-      <>
-        <footer>
-          <ul>
-            <li>Contact</li>
-            <li>About</li>
-            <li>Legal</li>
-          </ul>
-        </footer>
-      </>
-    </div>
-  );
+    );
 }
 
 export default App;
